derniere version de la GUI
git-svn-id: https://labeip.epitech.eu/svn/2011/livejamming/trunk@137 833c2cf1-1b31-df11-b905-005056bd6404

+++ src/server/Component_SessionManager.cpp
@@ -1,5 +1,6 @@
 #include <Component_SessionManager.h>
 
+
 int		timeOutTest_maxRetry = 3;
 
 Component_SessionManager::Component_SessionManager(IComponent::m_packet_bindings &packetBindings,
@@ -7,7 +8,7 @@ Component_SessionManager::Component_SessionManager(IComponent::m_packet_bindings
   :IComponent(serverManager), _packetBindings(packetBindings), _serverManager(serverManager)
 {
   _sessionMap = new m_Session;
-  _rng.seed(std::clock());
+  _rng.seed((int32_t)std::clock());
   PacketBindings();
 }
